+++ test/org/soulspace/qclojure/application/error_mitigation/zero_noise_test.clj
@@ -121,4 +121,8 @@
             (is (map? result) "Result should be a map")
             (is (contains? result :measurement-results) "Should contain measurement results")))))))
   
-
+(comment 
+  ;; Run all tests in namespace
+  (run-tests)
+  ;
+  )